Split gateset completeness tests
