+++ quri-parts/packages/circuit/tests/circuit/transpile/test_gateset.py
@@ -14,7 +14,6 @@ import numpy as np
 import pytest
 from numpy.typing import NDArray
 
-import quri_parts.circuit.transpile.rz2hst as rz2hst
 from quri_parts.circuit import (
     ImmutableQuantumCircuit,
     NonParametricQuantumCircuit,
@@ -102,44 +101,6 @@ def _assert_equivalent(
     assert abs(u1[idx]) > 1e-9
     phase = u2[idx] / u1[idx]
     assert np.allclose(u1 * phase, u2, atol=1e-6)
-
-
-# All pairs of single-qubit Clifford gate names that generate the full
-# single-qubit Clifford group (verified by group closure).
-_COMPLETE_SQ_CLIFFORD_GENERATORS: list[
-    tuple[CliffordGateNameType, CliffordGateNameType]
-] = [
-    (gate_names.H, gate_names.S),
-    (gate_names.H, gate_names.Sdag),
-    (gate_names.H, gate_names.SqrtX),
-    (gate_names.H, gate_names.SqrtXdag),
-    (gate_names.S, gate_names.SqrtX),
-    (gate_names.S, gate_names.SqrtXdag),
-    (gate_names.S, gate_names.SqrtY),
-    (gate_names.S, gate_names.SqrtYdag),
-    (gate_names.Sdag, gate_names.SqrtX),
-    (gate_names.Sdag, gate_names.SqrtXdag),
-    (gate_names.Sdag, gate_names.SqrtY),
-    (gate_names.Sdag, gate_names.SqrtYdag),
-    (gate_names.SqrtX, gate_names.SqrtY),
-    (gate_names.SqrtX, gate_names.SqrtYdag),
-    (gate_names.SqrtXdag, gate_names.SqrtY),
-    (gate_names.SqrtXdag, gate_names.SqrtYdag),
-]
-
-# Every single-qubit Clifford gate name.
-_ALL_SINGLE_QUBIT_CLIFFORD_GATES: list[GateNameType] = [
-    gate_names.H,
-    gate_names.X,
-    gate_names.Y,
-    gate_names.Z,
-    gate_names.S,
-    gate_names.Sdag,
-    gate_names.SqrtX,
-    gate_names.SqrtXdag,
-    gate_names.SqrtY,
-    gate_names.SqrtYdag,
-]
 
 
 def _single_qubit_clifford_circuit() -> ImmutableQuantumCircuit:
@@ -771,148 +732,3 @@ class TestGateSetConversionCliffordT:
         transpiled = GateSetConversionTranspiler(_CLIFFORD_T_TARGET)(circuit)
         assert _gate_kinds(transpiled) <= _CLIFFORD_T_GATE_NAMES
         assert gate_names.RZ not in _gate_kinds(transpiled)
-
-
-def _clifford_circuit() -> ImmutableQuantumCircuit:
-    circuit = QuantumCircuit(3)
-    circuit.extend(
-        [
-            gates.H(0),
-            gates.S(0),
-            gates.Sdag(1),
-            gates.X(1),
-            gates.Y(2),
-            gates.Z(0),
-            gates.SqrtX(1),
-            gates.SqrtXdag(2),
-            gates.SqrtY(0),
-            gates.SqrtYdag(1),
-            gates.CNOT(0, 1),
-            gates.CZ(1, 2),
-            gates.SWAP(0, 2),
-            gates.H(2),
-            gates.S(2),
-        ]
-    )
-    return circuit
-
-
-class TestCliffordConversionCompleteness:
-    # When the target gates form a complete set of single-qubit Clifford
-    # generators, CliffordConversionTranspiler converts EVERY single-qubit
-    # Clifford gate into the target gate set (and the conversion is correct).
-    @pytest.mark.parametrize("generators", _COMPLETE_SQ_CLIFFORD_GENERATORS)
-    def test_every_single_qubit_clifford_converts(
-        self, generators: tuple[CliffordGateNameType, CliffordGateNameType]
-    ) -> None:
-        for gate_name in _ALL_SINGLE_QUBIT_CLIFFORD_GATES:
-            circuit = QuantumCircuit(1)
-            circuit.add_gate(getattr(gates, gate_name)(0))
-            transpiled = CliffordConversionTranspiler(list(generators))(circuit)
-            assert _gate_kinds(transpiled) <= set(generators)
-            _assert_equivalent(circuit, transpiled)
-
-
-# Universal target gate sets needing no rotation approximation (gridsynth):
-# any non-MC circuit is transpilable into them exactly.
-_EXACT_UNIVERSAL_TARGETS: list[list[GateNameType]] = [
-    [gate_names.RX, gate_names.RY, gate_names.RZ, gate_names.CNOT],
-    [gate_names.RX, gate_names.RZ, gate_names.CNOT],
-    [gate_names.RY, gate_names.RZ, gate_names.CNOT],
-    [gate_names.RZ, gate_names.H, gate_names.CNOT],
-    [gate_names.RZ, gate_names.SqrtX, gate_names.X, gate_names.CNOT],
-    [gate_names.RX, gate_names.RY, gate_names.RZ, gate_names.CZ],
-    [gate_names.RX, gate_names.RY, gate_names.RZ, gate_names.SWAP, gate_names.CNOT],
-]
-
-
-class TestGateSetConversionCompleteness:
-    # Every (non-MC) input gate is convertible into a universal target gate set.
-    @pytest.mark.parametrize("target", _EXACT_UNIVERSAL_TARGETS)
-    def test_any_circuit_to_universal_set(self, target: list[GateNameType]) -> None:
-        circuit = _complex_circuit()
-        transpiled = GateSetConversionTranspiler(target)(circuit)
-        assert _gate_kinds(transpiled) <= set(target)
-        _assert_equivalent(circuit, transpiled)
-
-    # Every Clifford gate is converted to the output when the output is a
-    # complete set of Clifford generators (a complete single-qubit generator
-    # pair together with an entangler).
-    @pytest.mark.parametrize("entangler", [gate_names.CNOT, gate_names.CZ])
-    @pytest.mark.parametrize("generators", _COMPLETE_SQ_CLIFFORD_GENERATORS)
-    def test_clifford_circuit_to_complete_generators(
-        self,
-        generators: tuple[CliffordGateNameType, CliffordGateNameType],
-        entangler: GateNameType,
-    ) -> None:
-        target = [*generators, entangler]
-        circuit = _clifford_circuit()
-        transpiled = GateSetConversionTranspiler(target)(circuit)
-        assert _gate_kinds(transpiled) <= set(target)
-        _assert_equivalent(circuit, transpiled)
-
-    @pytest.mark.gridsynth
-    def test_complex_circuit_to_clifford_t_set(self) -> None:
-        # Clifford+T is universal too; arbitrary-angle rotations in the complex
-        # circuit are approximated into Clifford+T via gridsynth.
-        circuit = _complex_circuit()
-        transpiled = GateSetConversionTranspiler(_CLIFFORD_T_TARGET)(circuit)
-        assert _gate_kinds(transpiled) <= set(_CLIFFORD_T_TARGET)
-
-    def test_rotation_lowered_when_non_clifford_t_gate_in_target(self) -> None:
-        # The target contains TOFFOLI (not a Clifford+T gate), so it is not a
-        # subset of Clifford+T. Rotations must still be lowered into Clifford+T
-        # rather than left undecomposed, while TOFFOLI is kept.
-        target: list[GateNameType] = [
-            gate_names.H,
-            gate_names.S,
-            gate_names.T,
-            gate_names.CNOT,
-            gate_names.TOFFOLI,
-        ]
-        circuit = QuantumCircuit(3)
-        circuit.add_RZ_gate(0, np.pi / 4.0)
-        circuit.add_TOFFOLI_gate(0, 1, 2)
-        transpiled = GateSetConversionTranspiler(target)(circuit)
-        assert _gate_kinds(transpiled) <= set(target)
-        assert gate_names.RZ not in _gate_kinds(transpiled)
-        assert gate_names.TOFFOLI in _gate_kinds(transpiled)
-
-    @pytest.mark.parametrize(
-        "target",
-        [
-            [gate_names.S, gate_names.T, gate_names.X, gate_names.CNOT],
-            [gate_names.H, gate_names.T, gate_names.X, gate_names.CNOT],
-        ],
-    )
-    def test_rotation_not_lowered_when_gridsynth_output_is_not_supported(
-        self, target: list[GateNameType], monkeypatch: pytest.MonkeyPatch
-    ) -> None:
-        monkeypatch.setattr(
-            rz2hst,
-            "driver_pygridsynth",
-            lambda *args, **kwargs: pytest.fail("gridsynth should not be used"),
-        )
-        circuit = QuantumCircuit(1)
-        circuit.add_RZ_gate(0, 0.3)
-
-        with pytest.raises(ValueError, match="cannot be converted"):
-            GateSetConversionTranspiler(target)(circuit)
-
-    def test_rotation_in_target_but_t_not_in_target(self) -> None:
-        # A rotation gate (RZ) is in the target but T is not. T-like gates must
-        # be decomposed into the rotation (T is not kept), arbitrary-angle
-        # rotations are kept as RZ (no Clifford+T lowering / no gridsynth), and
-        # the result is exact.
-        target: list[GateNameType] = [gate_names.RZ, gate_names.H, gate_names.CNOT]
-        circuit = QuantumCircuit(2)
-        circuit.add_T_gate(0)
-        circuit.add_Tdag_gate(1)
-        circuit.add_RZ_gate(0, 0.3137)
-        circuit.add_H_gate(1)
-        circuit.add_CNOT_gate(0, 1)
-        transpiled = GateSetConversionTranspiler(target)(circuit)
-        assert _gate_kinds(transpiled) <= set(target)
-        assert gate_names.T not in _gate_kinds(transpiled)
-        assert gate_names.Tdag not in _gate_kinds(transpiled)
-        _assert_equivalent(circuit, transpiled)
